ml/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
pydantic==2.9.0
scikit-learn==1.5.2
numpy==1.26.4
joblib==1.4.2
pandas==2.2.3
python-dotenv==1.0.0
psycopg2-binary==2.9.10
supabase==2.4.0
httpx==0.27.0
